backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenv

pytz

pytz
supabase

livekit

openai
livekit-api


livekit-api
livekit

livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-noise-cancellation


python-dotenv
pytz

omegaconf
torch
torchaudio
numpy
scipy